normalizing package structure for clustering
explicitly exposed two functions in __init__
created a node module (former utils module), which is not imported by __init__ anymore.
update doc accordingly
